Add pagination and refactor reviews controller
I tried to improve performance removing iterations and unnecessary calls in the index method and delegate to fetch reviews from an AJAX request.

In the other hand I fix some security issues regarding SQL injection on the queries.

Finally to improve performance and avoid N+1 queries I tried to do only the necessary calls to the db and fetch product + reviews on 1 go.