Add RatingAnalyzer model
This class handles the average rating calculation for the last 3 months.
I decided to calculate everything directly in the db for performances reasons and then use a hash to present to the view with all the information needed.
I could also have refined a little more the presentation to the view but I think is ok for the scope of the exercise.